DEVEL15-pam-expires-setting-should-be-32bit-20071210
LICENSE IPL10

store as a 32 bit value explicitly so 32 bit routines don't leave garbage around here.


(cherry picked from commit c057063abea93733170c1b31c91b52050a150a98)

+++ src/pam/afs_setcred.c
@@ -62,7 +62,7 @@ pam_sm_setcred(pam_handle_t * pamh, int flags, int argc, const char **argv)
     int auth_ok = 0;
     char *lh;
     char *user = NULL;
-    long password_expires = -1;
+    int password_expires = -1;
     char *reason = NULL;
     struct passwd unix_pwd, *upwd = NULL;
     char upwd_buf[2048];	/* size is a guess. */
